Moved hardware stuff to another folder
