+++ EarTrumpet/DataModel/Hardware/HardwareConfiguration.cs
@@ -1,11 +1,11 @@
 ﻿using System.Xml.Serialization;
+using EarTrumpet.DataModel.MIDI;
 
-namespace EarTrumpet.DataModel.MIDI
+namespace EarTrumpet.DataModel.Hardware
 {
     [XmlInclude(typeof(MidiControlConfiguration))]
     public abstract class HardwareConfiguration
     {
         public abstract override string ToString();
-        public abstract void FromString(string str);
     }
 }
